Create costs module
